Set up the add student view and the tags for it.
As needed for issue #3, I added a view controller for adding a new student.
I also set up the basics of a table view in that add student view that will allow the addition of tags for the students. This allows the user to add different tags to a student, such as the class they are in, their phone number, etc.