Added the ability to select the tag type.
The tag popover appears with a list of possible tags. It is populated and won't scroll if everything fits within the bounds. Selecting a tag type affects the text field and will change the image presented to the user, as well as the placeholder text.

This is a major step toward completing issue #3.